docker/requirements.txt
tensorflow-cpu
pandas
numpy
scikit-learn
joblib
